Consolidate JointWrench sensor tests
